feat: add delete image and album from cloudinary

+++ routes/albumRoutes.js
@@ -1,6 +1,10 @@
 const express = require('express');
 const router = express.Router();
 const Image_album = require('../model/image_album.model.js');
+const Image = require('../model/image.model.js');
+const cloudinary = require('../cloudinaryConfig');
+
+const { getPublicIdFromUrl } = require('../utils/getPublicIdFromUrl.js');
 
 
 // Отримати всі альбоми
@@ -75,14 +79,45 @@ router.put('/update-album', async (req, res) => {
   }
 });
 
-// Видалити альбом
+
 router.post('/delete-album', async (req, res) => {
   const albumId = req.query.albumId;
+
   try {
-    const deletedAlbum = await Image_album.findByIdAndDelete(albumId );
-    res.status(200).json(deletedAlbum);
+    if (!albumId) {
+      return res.status(400).json({ message: "albumId is required" });
+    }
+
+    // Знайти всі зображення альбому
+		const images = await Image.find({album_id: albumId });
+	console.log(images);
+	
+		
+
+    // Видалити з Cloudinary усі зображення альбому
+    for (const img of images) {
+			const publicId = getPublicIdFromUrl(img.img);
+			console.log(publicId);
+			
+      if (publicId) {
+        await cloudinary.uploader.destroy(publicId);
+      }
+    }
+
+    // Видалити всі зображення з MongoDB
+    await Image.deleteMany({ album_id: albumId });
+
+    // Видалити альбом
+    const deletedAlbum = await Image_album.findByIdAndDelete(albumId);
+
+    if (!deletedAlbum) {
+      return res.status(404).json({ message: "Album not found" });
+    }
+
+    res.status(200).json({ message: "Album and related images deleted", album: deletedAlbum });
   } catch (error) {
-    res.status(400).json({message : error.message});
+    console.error("Delete album error:", error);
+    res.status(500).json({ message: error.message });
   }
 });
